Created new folders
Using a class inheritance structure where a class is a folder and inheriting classes are child folders. "Resources" are created and stored in their respective class folder, for example "res://texture/2d/compressed/icon.svg" is an SVG icon "Resource" that is a "CompressedTexture2D" -> "Texture2D" -> "Texture" -> "Resource". the resource class also inherits from "Refcounted" -> "Object" but since the Godot project structure is based on "Resource" files we do not need parent folders for those classes, with the exception of script files which may be a simple "Object" or "RefCounted" script with custom functionality.